Create custom_integrations/ecowitt_iot
An exact copy of core_integrations/ecowitt contents as of 2024-11-10.

Created as the existing core ecowitt integration is a push based implementation, and the Ecowitt IOT devices need a polling/local api approach. Smooshing them together seems overly complicated, so github.com/Bwooce/ecowitt_iot is my new integration to do this in a standalone integration. There's no cross-over interference with the other integration.

Duplicating files is not my preferred approach but symlinking is contrary to the guidance that custom_integrations may not symlink (for security reasons, I'd assume).