Add icons for custom comelit_intercom integration
The official `comelit` integration so far only works for their "Comelit
SimpleHome" product. However, the company also offers a video intercom
system that is WiFi connected, and many users were looking for a way of
integrating this with Home Assistant:
https://community.home-assistant.io/t/comelit-video-intercom/15128/78

I have recently finally release a first public 0.1.0 version of such a
custom integration:
https://github.com/nicolas-fricke/ha-component-comelit-intercom

To not collide and overwrite the SimpleHome integration, it is now
called `comelit_intercom`, but the logos should be the same as for the
official Comelit integration. Therefore, copy the logos over to a folder
with the name of the new integration.